refactor: move kycc redirect to marketplace component
